refactor: engine ray casting
