+++ .vscode/settings.json
@@ -1,5 +1,7 @@
 {
     "files.associations": {
-        "cub3d.h": "c"
+        "cub3d.h": "c",
+        "structs.h": "c",
+        "libft.h": "c"
     }
 }
